Moved funcs from main on gamemode.cpp, included game loop
